+++ R/pin_ll_CAR.R
@@ -7,12 +7,12 @@
 #' \code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'} (in this order).
 #' Vectors for \code{numbuys} and \code{numsells} need to have same length.
 #'
-#' @param param \emph{numeric}: (named) vector of model parameters
+#' @param param \emph{numeric} (named) vector of model parameters
 #'                              (valid names: \code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'}),
 #'                              length must equal 5
 #' @param numbuys \emph{numeric} vector of daily buys
 #' @param numsells \emph{numeric} vector of daily sells
-#' @param factorization \emph{character}: switch between EHO ('EHO') and Lin-Ke ('Lin_Ke') factorization
+#' @param factorization \emph{character} switch between EHO ('EHO') and Lin-Ke ('Lin_Ke') factorization
 #'
 #' @return \emph{numeric}: likelihood function value
 #'